% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upset.R
\name{prepare_upset_matrix}
\alias{prepare_upset_matrix}
\title{Prepare matrix data for upset plot}
\usage{
prepare_upset_matrix(df, item_col, set_col)
}
\arguments{
\item{df}{Input data frame}

\item{item_col}{Name of the item column}

\item{set_col}{Name of the set column}
}
\value{
A list with `matrix` and `segment` data frames
}
\description{
Internal function to prepare the data structure needed for upset plotting
}
\keyword{internal}
