% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_hrange.R
\name{annotate_hrange}
\alias{annotate_hrange}
\title{Title}
\usage{
annotate_hrange(
  xmin,
  xmax,
  y,
  scale = 1,
  label = NULL,
  tip_length = 0,
  line.size = 0.2,
  text.size = 2
)
}
\arguments{
\item{text.size}{}
}
\value{

}
\description{
Title
}
