% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distance_to_gene.R
\name{compute_distance_to_gene}
\alias{compute_distance_to_gene}
\title{Title}
\usage{
compute_distance_to_gene(
  txdb,
  chromosome,
  start,
  end,
  ref_position,
  type = c("GB", "TSS")
)
}
\arguments{
\item{type}{}
}
\value{

}
\description{
Title
}
