% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cs_color_mapping.R
\name{get_cs_color_mapping}
\alias{get_cs_color_mapping}
\title{Title}
\usage{
get_cs_color_mapping(
  cs_ids,
  highlight_cs_ids = NULL,
  colors = BuenColors::jdb_palette("corona")[setdiff(seq(15), c(8, 15))]
)
}
\arguments{
\item{colors}{}
}
\value{

}
\description{
Title
}
