% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_theme.R
\name{get_default_theme}
\alias{get_default_theme}
\title{Title}
\usage{
get_default_theme(
  fontsize = 8,
  legend.position = c(1, 1),
  legend.justification = c(1, 1),
  hide.xlab = FALSE,
  hide.ylab = FALSE,
  hide.xtext = FALSE,
  hide.ytext = FALSE,
  hide.xtitle = FALSE,
  hide.ytitle = FALSE
)
}
\arguments{
\item{hide.ytitle}{}
}
\value{

}
\description{
Title
}
