% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_global_position.R
\name{get_global_position}
\alias{get_global_position}
\title{Title}
\usage{
get_global_position(chromosome, position, reference_genome)
}
\arguments{
\item{reference_genome}{}
}
\value{

}
\description{
Title
}
