% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fm_panel.R
\name{plot_fm_panel}
\alias{plot_fm_panel}
\title{Title}
\usage{
plot_fm_panel(
  data,
  highlight_pos = NULL,
  title = NULL,
  legend_title = "95\% CS",
  xlim = NULL,
  ylim = c(0, 1),
  ybreaks = seq(0, 1, by = 0.2),
  point.size = 1.5,
  point.size2 = 3,
  ggtheme = get_default_theme(),
  rasterize = FALSE,
  rasterize.dpi = 300,
  relevel.cs_id = TRUE
)
}
\arguments{
\item{ggtheme}{}
}
\value{

}
\description{
Title
}
