% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_panel.R
\name{plot_gene_panel}
\alias{plot_gene_panel}
\title{Title}
\usage{
plot_gene_panel(
  chromosome,
  start,
  end,
  genome_build = c("hg19", "hg38"),
  txdb = NULL,
  highlight_pos = NULL,
  gene_col = BuenColors::jdb_palette("calma_azules")[6],
  fontsize = 8,
  point.size = 2,
  label.size = 2,
  arrow.rate = 0.015,
  length = unit(0.1, "cm"),
  background.layers = NULL
)
}
\arguments{
\item{label.size}{}
}
\value{

}
\description{
Title
}
