% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locuszoom.R
\name{plot_locuszoom}
\alias{plot_locuszoom}
\title{Title}
\usage{
plot_locuszoom(
  data,
  highlight_pos = NULL,
  window = NULL,
  xlim = NULL,
  manhattan.args = list(),
  manhattan.title = NULL,
  manhattan.breaks = ggplot2::waiver(),
  manhattan.loglog_p = TRUE,
  nlog10p_threshold = 0,
  fm.args = list(),
  fm.ylim = c(0, 1),
  fm.breaks = seq(0, 1, by = 0.2),
  fm.legend_title = "95\% CS",
  r2.args = list(),
  gene.args = list(),
  gene_score.args = list(),
  plot.manhattan = TRUE,
  plot.fm = TRUE,
  plot.r2 = FALSE,
  plot.gene = TRUE,
  plot.gene_score = FALSE,
  fontsize = 8,
  patchwork = TRUE,
  rasterize = FALSE,
  rasterize.dpi = 300
)
}
\arguments{
\item{plot.gene_score}{}
}
\value{

}
\description{
Title
}
