% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_manhattan_panel.R
\name{plot_manhattan_panel}
\alias{plot_manhattan_panel}
\title{Title}
\usage{
plot_manhattan_panel(
  data,
  highlight_pos = NULL,
  xlim = NULL,
  ylim = NULL,
  ybreaks = ggplot2::waiver(),
  nlog10p_threshold = 1,
  loglog_p = 10,
  plot.loglog_p = FALSE,
  point.size = 1.5,
  point.size2 = 3,
  line.size = 0.5,
  title = NULL,
  r2_cols = c("navy", "lightskyblue", "green", "orange", "red"),
  lead_variant_col = "purple3",
  ggtheme = get_default_theme(),
  background.layers = NULL,
  rasterize = FALSE,
  rasterize.dpi = 300
)
}
\arguments{
\item{lead_variant_col}{}
}
\value{

}
\description{
Title
}
