% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_r2_panel.R
\name{plot_r2_panel}
\alias{plot_r2_panel}
\title{Title}
\usage{
plot_r2_panel(
  data,
  highlight_pos = NULL,
  xlim = NULL,
  ylim = c(0, 1),
  ybreaks = seq(0, 1, by = 0.2),
  nlog10p_threshold = 1,
  ggtheme = get_default_theme(),
  rasterize = FALSE,
  rasterize.dpi = 300
)
}
\arguments{
\item{rasterize.dpi}{}
}
\value{

}
\description{
Title
}
