% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Title}
\usage{
preprocess(
  data,
  lead_variant = NULL,
  variant_col = "variant",
  beta_col = "beta",
  se_col = "se",
  pvalue_col = "pvalue",
  pip_col = "pip",
  cs_id_col = "cs_id",
  r2_col = "r2"
)
}
\arguments{
\item{r2_col}{}
}
\value{

}
\description{
Title
}
