% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_chromosome.R
\name{scale_color_chromosome}
\alias{scale_color_chromosome}
\title{Title}
\usage{
scale_color_chromosome(odd_color = "darkblue", even_color = "grey50")
}
\arguments{
\item{even_color}{}
}
\value{

}
\description{
Title
}
