% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpSet2.R
\name{UpSet2}
\alias{UpSet2}
\title{Modified UpSet plot}
\usage{
UpSet2(
  m,
  set_on_rows = TRUE,
  comb_col = "black",
  pt_size = grid::unit(3, "mm"),
  lwd = 2,
  bg_col = "#F0F0F0",
  bg_pt_col = "#CCCCCC",
  set_order = NULL,
  comb_order = NULL,
  top_annotation = NULL,
  right_annotation = NULL,
  row_names_side = "left",
  remove_lines = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Modified UpSet plot
}
